% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.effects.R
\name{desc.effects}
\alias{desc.effects}
\title{Describe the effects}
\usage{
desc.effects(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{list, optional
Additional arguments.}
}
\value{
{Effects, standard errors and confidence intervals of an object}
}
\description{
Describe the effects, and calculate standard
errors and confidence intervals
}
\examples{
\dontrun{
# describe the effects of cig_med, a mediation object
desc.effects(cig_med)
}
}
\seealso{
\code{\link{desc.effects.mediation}}, \code{\link{wgtmed}}
}
