% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSDUH_female}
\alias{NSDUH_female}
\title{A dataset containing the substance use condition and sexual orientation
of 40293 women.
.}
\format{
A data frame with 40293 rows and 24 variables:
\describe{
\item{lgb_flag}{whether or not belong to LGB, binary variable}
\item{cig15}{early smoking initiating at 15 years old}
\item{cigmon}{current smoker when taking the survey}
...
}
}
\source{
\url{https://nsduhweb.rti.org/respweb/homepage.cfm}
}
\usage{
NSDUH_female
}
\description{
A dataset containing the substance use condition and sexual orientation
of 40293 women.
.
}
\keyword{datasets}
