% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bal.table.mediation.R
\name{print.bal.table.mediation}
\alias{print.bal.table.mediation}
\title{Default print statement for \code{mediation} class}
\usage{
\method{print}{bal.table.mediation}(x, ...)
}
\arguments{
\item{x}{A \code{bal.table.mediation} object.}

\item{...}{Additional arguments.}
}
\value{
Default print statement
}
\description{
Default print statement for \code{mediation} class
}
\examples{
\dontrun{
print(bal.table.mediation(cig_med))
}
}
\seealso{
\code{\link{bal.table.mediation}}, \code{\link{wgtmed}}
}
