% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mediation.R
\name{summary.mediation}
\alias{summary.mediation}
\title{Displays a useful description of a \code{mediation} object.}
\usage{
\method{summary}{mediation}(object, ...)
}
\arguments{
\item{object}{A \code{mediation} object}

\item{...}{Additional arguments.}
}
\value{
\item{ps_tables}{Table of observations' propensity scores}

\item{mediator_distribution_check}{balance tables for NIE_1 and NIE_0}
}
\description{
Displays a useful description of a \code{mediation} object.
}
\examples{
\dontrun{
# Summary a mediation object
summary(cig_med) 
}
}
\seealso{
\code{\link{wgtmed}}
}
