% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mediation.R
\name{swapTxCtrl}
\alias{swapTxCtrl}
\title{Auxiliary function to swap treatment and control in one element of a desc
object from a ps object or dx.wts object}
\usage{
swapTxCtrl(dd)
}
\arguments{
\item{dd}{numeric
An element of a desc object from a ps or dx.wts object}
}
\value{
A desc object with swapped treatment and control
}
\description{
Call this in the \code{wgtmed()} function and the \code{bal.table.mediation()} function.
}
\seealso{
\code{\link[twang]{ps}}, \code{\link{dx.wts.mediation}}
}
