% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twangMediation-package.R
\docType{package}
\name{twangMediation-package}
\alias{twangMediation}
\alias{twangMediation-package}
\title{twangMediation: Twang Causal Mediation Modeling via Weighting}
\description{
Provides functions for estimating natural direct and indirect effects for mediation analysis. It uses weighting where the weights are functions of estimates of the probability of exposure or treatment assignment (Hong, G (2010). <https://cepa.stanford.edu/sites/default/files/workshops/GH_JSM%20Proceedings%202010.pdf> Huber, M. (2014). <doi:10.1002/jae.2341>). Estimation of probabilities can use generalized boosting or logistic regression. Additional functions provide diagnostics of the model fit and weights. The vignette provides details and examples.
}
\author{
\strong{Maintainer}: Dan McCaffrey \email{dmccaffrey@ets.org}

Authors:
\itemize{
  \item Katherine Castellano \email{kecastellano@ets.org}
  \item Donna Coffman \email{donna.coffman@temple.edu}
  \item Brian Vegetabile \email{bvegetab@rand.org}
  \item Megan Schuler \email{mschuler@rand.org}
  \item Haoyu Zhou \email{haoyu.zhou@temple.edu}
}

}
\keyword{propensity}
\keyword{score}
