% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.effects.R
\name{desc.effects}
\alias{desc.effects}
\title{Describe the effects}
\usage{
desc.effects(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{list, optional
Additional arguments.}
}
\value{
{Effects, standard errors and confidence intervals of an object}
}
\description{
Describe the effects, and calculate standard
errors and confidence intervals
}
\examples{
\dontrun{
#' data(NSDUH_female)
TEps <- ps(formula = lgb_flag ~ age + race + educ + income + employ,
data=NSDUH_female, verbose=F, n.trees=6000, estimand="ATE", stop.method="ks.mean")
# Obtain mediaiton estimates of interest
cig_med <- wgtmed(formula.med = cig15 ~ age + race + educ + income + employ,
   a_treatment="lgb_flag",
   y_outcome="cigmon",
   data=NSDUH_female,
   method="ps",
   total_effect_ps=TEps,
   total_effect_stop_rule="ks.mean",
   ps_version="gbm",
   ps_n.trees=6000,
   ps_interaction.depth=3,
   ps_shrinkage=0.01,
   ps_stop.method="ks.mean",
   ps_verbose=FALSE)
# describe the effects, standard errors, and confidence intervals
desc.effects(cig_med)
}
}
