% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSDUH_female}
\alias{NSDUH_female}
\title{A dataset containing the substance use condition and sexual orientation
of 40293 women.
.}
\format{
A data frame with 40293 rows and 24 variables:
\describe{
\item{cigmon}{Past month daily cigerretes use, yes or no}
\item{lgb_flag}{whether or not belong to LGB, binary variable}
\item{cig15}{early smoking initiating at 15 years old, yes or no}
\item{alc15}{early alcohol use initiating at 15 years old}
\item{educ}(education level, four categorical levels)
\item{income}{income level, four categorical levels}
\item{race}{race, four categories}
\item{employ}{employment, yes or no}
\item{alc_cig_depend}{nicotine or alcohol dependence, yes or no}
...
}
}
\source{
\url{https://nsduhweb.rti.org/respweb/homepage.cfm}
}
\usage{
NSDUH_female
}
\description{
A dataset containing the substance use condition and sexual orientation
of 40293 women.
.
}
\keyword{datasets}
