% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mediation.R
\name{print.mediation}
\alias{print.mediation}
\title{Default print statement for \code{mediation} class}
\usage{
\method{print}{mediation}(x, ...)
}
\arguments{
\item{x}{A \code{mediation} object.}

\item{...}{Additional arguments.}
}
\value{
Default print statement.
}
\description{
Default print statement for \code{mediation} class
}
\examples{
\dontrun{
data(NSDUH_female)
TEps <- ps(formula = lgb_flag ~ age + race + educ + income + employ,
data=NSDUH_female, verbose=F, n.trees=6000, estimand="ATE", stop.method="ks.mean")
# Obtain mediaiton estimates of interest
cig_med <- wgtmed(formula.med = cig15 ~ age + race + educ + income + employ,
   a_treatment="lgb_flag",
   y_outcome="cigmon",
   data=NSDUH_female,
   method="ps",
   total_effect_ps=TEps,
   total_effect_stop_rule="ks.mean",
   ps_version="gbm",
   ps_n.trees=6000,
   ps_interaction.depth=3,
   ps_shrinkage=0.01,
   ps_stop.method="ks.mean",
   ps_verbose=FALSE)
# Now print
print(cig_med)
}
}
