% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mediation.R
\name{summary.mediation}
\alias{summary.mediation}
\title{Displays a useful description of a \code{mediation} object.}
\usage{
\method{summary}{mediation}(object, ...)
}
\arguments{
\item{object}{A \code{mediation} object}

\item{...}{Additional arguments.}
}
\value{
\item{ps_tables}{Table of observations' propensity scores}

\item{mediator_distribution_check}{balance tables for NIE_1 and NIE_0}
}
\description{
Displays a useful description of a \code{mediation} object.
}
\examples{
\dontrun{
data(NSDUH_female)
# Calculate total effect weighs
TEps <- ps(formula = lgb_flag ~ age + race + educ + income + employ,
data=NSDUH_female, verbose=F, n.trees=6000, estimand="ATE", stop.method="ks.mean")
# Obtain mediaiton estimates of interest
cig_med <- wgtmed(formula.med = cig15 ~ age + race + educ + income + employ,
   a_treatment="lgb_flag",
   y_outcome="cigmon",
   data=NSDUH_female,
   method="ps",
   total_effect_ps=TEps,
   total_effect_stop_rule="ks.mean",
   ps_version="gbm",
   ps_n.trees=6000,
   ps_interaction.depth=3,
   ps_shrinkage=0.01,
   ps_stop.method="ks.mean",
   ps_verbose=FALSE)
# Summary cig_med
summary(cig_med)
}
}
