% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mediation.R
\name{summary.mediation}
\alias{summary.mediation}
\title{Displays a useful description of a `mediation` object.}
\usage{
\method{summary}{mediation}(object, ...)
}
\arguments{
\item{object}{A `mediation` object}

\item{...}{Additional arguments.}
}
\description{
Displays a useful description of a `mediation` object.
}
